extract PhotoGrid class
